refactor: Kebab case DetectAdBlocker
