+++ src/detect-ad-blocker.ts
